applespi: Fix initialization failures and keyboard slowness.
There were two issues here. First was the missing delays after each
transaction. The value used here (100us) was derived from the
'spiCSDelay' value found in the DSDT (the value given there is 10, but
with no units - experimentation found that 100us was the minimum value
at which things were stable, so assuming the unit is 10us). This then
removed the need for reducing the clock frequency, which in turn appears
to fix slow keyboard issues on MacBook8,1 (2015 model).

The second issue was an extra cs-change after sending the init message
and before receiving the status. The code comment was actually correct,
but the code didn't do what the comment said.

Lastly, this introduces a separate buffer to receive the status response
from the write.