applespi: Introduce delay between cs and receive on reads.
On MacBook9,1 this delay appears to be necessary, otherwise the reads
include leading garbage.

There are two unknown items in the settings returned by the _DSM method
in the DSDT. Both specify a value of 10µs, and a delay with that value
solves the above read issues. I'm guessing resetA2R stands for
active-to-receive, and hence picking that one to use here.