applespi: Remove the need to modify the DSDT.
Much of the information supplied in the DSDT resource-template returned
when the DSDT is modified is also available from the _DSM method. So we
register ourselves as an ACPI driver for the device, and then create and
register the spi device ourselves, configuring it from the information
returned by the _DSM method.