applespi: Change the default for iso-layout to disabled.
Unlike the hid-apple driver, we don't have any idea whether the current
keyboard is using the iso-layout or not. This means that the iso_layout
has subtly different semantics here than on hid-apple (from which this
was copied): rather than just allowing for disabling of the iso-layout
on iso keyboards, the flag here determines which layout is used on all
keyboards. Therefore the flag's default should reflect the most common
keyboard type, which is not iso.