applespi: Make master-spi-device detection more robust.
The previous implementation was assuming that the uid of acpi-device
associated with the spi-master device was the same as the bus-number
of that device. While this was indeed the case on MBP's, at least on
some MB's it isn't.

Instead we now look at the companion acpi-devices of the spi-master
and slave devices, and look for the master that matches the slave's
parent. This should work correctly everywhere.