applespi: Fix race during spi-master registration.
The spi-master-device registration registers the device before it starts
the queue and marks the device as running. Because device registration
immediately runs all class hooks, we end up with a small window where we
are notified that the spi-master device is available but it's not
actually ready to receive commands. Therefore we now check if spi master
is marked running and wait a bit and retry if not.

This fixes some sporadic touchpad ininitialization failures.