applespi: Reorder init code to avoid races.
In particular we need to ensure the keyboard input handler is registered
after all structures are initialized because it registers an event
handler and hence we can be called at any point after registration.

Also moved the USB check to the very top, since there's no point in
doing anything if that succeeds.