applespi: Initialize spi message and transfers only once.
It is apparently safe to initialize these once and keep re-using them.
A number of existing drivers do this already.

Also inlined applespi_setup_spi_message() since it's only used in two
places and didn't seem to be worth it.

This is in response to a review by Lukas Wunner.