applespi: Refine the packet data structures.
More fields in the spi packet(s) have been identified, and the
structures reworked accordingly. In particular several fields in the
overall header of the spi packet, as well the message header have been
clarified. However, a number of fields are still unknown or have "magic"
values; and perforce there's some random guesswork such as whether
something is two 1-byte fields or a single 2-byte field.

This lays the groundwork for improving the packet handling in the next
commits.