applespi: Don't set up any runtime-pm callbacks.
We aren't doing any runtime power management (and don't need to), and
even if we were it's unlikely our current suspend/resume callbacks
would be appropriate.

Furthermore, since our suspend/resume callbacks are not for runtime
power management, they don't need to (and should not) be inside
CONFIG_PM ifdef's.