applespi: Avoid unnecessary delay in common cases.
Most operations are reads (i.e. keyboard and touchpad events);
furthermore the read-write-change delay is only necessary when actually
switching between read and write operations. So putting the delay after
each read is in most cases unnecessary. Hence this moves the delay to
the start of the writes instead (note that because each write is
followed by a read, it follows that each write is preceded by a read,
and so the delay at the start of the write is always necessary).