appletb: Fix double registration of input devices etc.
Our probe gets called for both the simple and the fancy USB interfaces,
so we need to only fully initialize for the former. This fixes USB
errors due to the interfaces being accessed twice. And it also avoids
confusion with the parameters being available on multiple devices, but
only being honored on the simple device.