appletb: Add support for dynamic ALS sensitivity.
MacOS manages to configure the sensor so it only reports changes when
the value changes by more than 15% of the current value. Since I haven't
been able to figure out how it does that so far, this implements a poor
man's version of that by dynamically changing the sensitivity depending
on the current brightness value.