appletb: Split the module into 3 separate ones for iBridge, TB, and ALS.
With the inclusion of the ALS functionality the single module has become
unwieldy and badly named. So the code was split into an apple-ibridge
mfd driver, an apple-ib-tb hid driver, and an apple-ib-als iio driver.

The apple-ibridge driver contains the common, iBridge wide code, as well
as some helpers. The apple-ib-tb now contains just the touchbar
functionality, and the apple-ib-als just the ALS functionality.