appletb: Ensure tb device is powered on before changing mode.
Usually this is already the case, as we disable autopm whenever the
display is on. However, if the display is off, and the mode is changed
(i.e. when you press the fn key after the display has gone idle) then
the device could be powered down, resulting in an error when trying to
set the mode.

This fixes #15.