appletb: Multiple fixes for suspend/resume.
Need to ensure the TB device is in a safe state when suspending, or it
might end up in a weird state on resume. Furthermore, the autopm state
is not preserved, so that needs to be taken care of.

Additionally, suppress changing the mode or display while the device is
suspended - this mainly avoids ugly error messages in the log.