appletb: Avoid flicker when turning on touchbar again.
In the case where the fn key triggers the touchbar to wake up again, the
display is often turned on slightly before the mode change takes
efffect, leading to a sort of a "flicker" as the mode is changed. So we
introduce a small delay between the mode and disp changes in case both
change (which usually just happens in this wakeup case).