appletb: Turn 64-bit hid report fields into 32-bit ones.
The spec says fields can't be more than 4 bytes, and hid_extract_field()
complains when encountering such fields. Furthermore, because one of
them is in an input report that is received every 40 minutes, we get a
number of ugly error messages in the kernel log. Turning the fields into
32-bit ones avoids this.

This fixes #16.