applespi: split the 16-bit model-id into 2 8-bit fields.
We don't exactly know the significance of these 2 bytes, other than they
1) are the only 2 bytes that vary between different trackpads
2) their values are fixed for a given trackpad size

Looking at their values, we note that the first byte has values 4, 5,
and 6, and the second byte has values 0x17, 0x57, and 0xd7. The latter
exhibits the fact that each one is a superset of the previous in terms
of bits set, which hints that these might be flags. On top of this, the
two bytes as a 16-bit value are not naturally aligned in the struct (and
hence necessitating the __packed) - all other fields in all other
structs are aligned, however.

For these reasons this splits up the 16-bit model_id into 2 8-byte
fields, model_no and model_flags, thereby avoiding the need for
__packed. The trackpad info is now looked up based on just the model_no,
but product_id in the input device is left as a combination of these two
fields for better specificity.

From the review by Andy Shevchenko.