appletb: fix passing of ib_dev instance to hid probe callback.
The existing solution was an ugly hack: a static appleib_dev variable
was used to pass the ib_dev pointer from the acpi-probe callback to the
hid-driver probe callback. Instead we now pass this pointer via the
driver_data field of the hid_device_id. This necessitates making a copy
of the hid_driver and hid_device_id structures, though, to hold the
modified data.

From the review by Jonathan Cameron.